Restructured the packages.
